Fixed clang warnings
- Modified the cmake project file as well to reduce warnings further.

+++ include/spdlog/details/os-inl.h
@@ -358,7 +358,7 @@ SPDLOG_INLINE size_t thread_id() SPDLOG_NOEXCEPT
 SPDLOG_INLINE void sleep_for_millis(int milliseconds) SPDLOG_NOEXCEPT
 {
 #if defined(_WIN32)
-    ::Sleep(milliseconds);
+    ::Sleep(static_cast<DWORD>(milliseconds));
 #else
     std::this_thread::sleep_for(std::chrono::milliseconds(milliseconds));
 #endif
